WIP: Relax SG rules for IP fragmentation (L4 rules)
For SG rules which need to examine L4 fields,
allow a fragmented packet if:

- it's the first fragment (offset == 0) and it's big enough to include
  the complete L4 header, and the header is matching the rule,
- or it's a non-first fragment (offset > 0) and its L3 header is matching
  the rule.

The second case might be controversial, as it allows packets which
doesn't actually match with the original SG rule.  This is a
loose-but-efficient approach.  A more strict alternative would involve
IP defragmentation before applying rules, which might or might not be
a good idea.

Ref: MI-573
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I479b7d045651777991bb66671b1eb1ff765fc7dd