MeterRegistry: Fix metering
After the recent change, [1] the FlowMatch instance given to
trackFlow() can be reused for an unrelated flow while our HashMap
still has a reference to it.  It broke metering as updateFlow() fails
to update the correct meters.  It also leads memory leaks because the
corresponding forgetFlow() will fail to remove the key from the
HashMap.  This commit fixes the problems by making a copy of the key.

Alternatively, we can make removeFlowFromDatapath swap FlowMatch
instances in FlowOperation and ManagedFlowImpl.  That way it would
produce less garbage.  However, it complicates interfaces between
modules.  Anyway, if you seriously care the amount of garbage, the
"native" implementation would be a better choice these days.

[1] Ib1cc97cfdf07d3d6973116f9d07860355aaedd3e
    ("Remove ref counting from flow operation")

Ref: MI-2677
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: Ie8d2408428a789ba29452b62543319af599e55b4